+++ app/DI/RepositoriosConfig.cs
@@ -10,7 +10,6 @@ namespace app.DI
             services.AddScoped<IUnidadeFederativaRepositorio, UnidadeFederativaRepositorio>();
             services.AddScoped<IUsuarioRepositorio, UsuarioRepositorio>();
             services.AddScoped<IPerfilRepositorio, PerfilRepositorio>();
-            services.AddScoped<IPermissaoRepositorio, PermissaoRepositorio>();
         }
     }
 }
