+++ app/DI/ServicesConfig.cs
@@ -18,6 +18,7 @@ namespace app.DI
             services.AddScoped<AuthService>();
             services.AddScoped<IEmailService, EmailService>();
             services.AddScoped<IPerfilService, PerfilService>();
+            services.AddScoped<IPermissaoService, PermissaoService>();
 
             services.AddAuth(configuration);
         }
