--- app/Repositorios/Interfaces/IPermissaoRepositorio.cs
@@ -1,10 +1,0 @@
-using api;
-
-namespace app.Repositorios.Interfaces
-{
-    public interface IPermissaoRepositorio
-    {
-        public List<string> ObterCategorias();
-        public List<string[]> ObterPermissoesPortCategoria(string categoria);
-    }
-}
