+++ app/Services/Interfaces/IPermissaoService.cs
@@ -1,0 +1,10 @@
+using api.Permissoes;
+
+namespace app.Services.Interfaces
+{
+    public interface IPermissaoService
+    {
+        public List<string> ObterCategorias();
+        public List<PermissaoModel> ObterPermissoesPortCategoria(string categoria);
+    }
+}
