+++ test/Fixtures/Base.cs
@@ -27,12 +27,12 @@ namespace test.Fixtures
             services.AddScoped<IUnidadeFederativaRepositorio, UnidadeFederativaRepositorio>();
             services.AddScoped<IUsuarioRepositorio, UsuarioRepositorio>();
             services.AddScoped<IPerfilRepositorio, PerfilRepositorio>();
-            services.AddScoped<IPermissaoRepositorio, PermissaoRepositorio>();
 
             // Services
             services.AddScoped<IUsuarioService, UsuarioService>();
             services.AddScoped<IEmailService, EmailService>();
             services.AddAutoMapper(typeof(AutoMapperConfig));
+            services.AddScoped<IPermissaoService, PermissaoService>();
 
             // Controllers
             services.AddScoped<DominioController>();
